Adds a connection timeout to the Postgres pool
At the moment the filler is _very_ slow, only processing some blocks
every few minutes (sometimes even being forced to restart after 10
minutes of no block processed).

I have a feeling that it might have to do with deadlocks due to
postgres connections being exhausted (but no proof, and it would be
weird if this is a new issue). To have better knowledge on that, adds
a connection timeout so we can see errors if anything fails due to no
connection being available.