Disables call to update_atomicmarket_sales_filters
The filler is still quite slow. There are no errors regarding database
connection timeouts, so that was not the issue. The function
update_atomicmarket_sales_filters comes up as a very slow function in
the RDS performance reporting though. It _seems_ that we might be able
to call this later once we have caught up, so for now we disable it.